requirements.txt
djangorestframework
psycopg2
firebase-admin
pillow
corsheaders